Release 1.0

+++ app/Http/Services/Media/UpdateMediaService.php
@@ -2,7 +2,10 @@
 
 namespace App\Http\Services\Media;
 
+use App\Repositories\AnimalMediaRepository;
+use App\Repositories\EventMediaRepository;
 use App\Repositories\MediaRepository;
+use App\Repositories\UserRepository;
 use Illuminate\Support\Facades\Storage;
 
 class UpdateMediaService
@@ -13,14 +16,37 @@ class UpdateMediaService
 
         $media = $repository->getById($id);
 
-        $oldFileName = $media->filename;
+        $isCover = data_get($data, 'is_cover');
 
-        $data = (new CreateMediaService())->makeUpload($data);
+        if ($isCover !== null) {
+            $normalized = $this->normalize(data_get($data, 'origin'));
+
+            if ($normalized !== UserRepository::class) {
+                $normalized->newQuery()->where('media_id', $media->id)->update(['is_cover' => $isCover]);
+            }
+        }
+
+        if (data_get($data, 'media')) {
+            $oldFileName = $media->filename;
+
+            $data = (new CreateMediaService())->makeUpload($data);
+
+            Storage::disk('google')->delete($oldFileName);
+        }
 
         $updated = $repository->update($media, $data);
 
-        Storage::disk('google')->delete($oldFileName);
-
         return $updated;
+    }
+
+    private function normalize(string $origin)
+    {
+        $normalized =  [
+            'event' => new EventMediaRepository(),
+            'animal' => new AnimalMediaRepository(),
+            'user' => new UserRepository(),
+        ];
+
+        return $normalized[$origin];
     }
 }
